+++ History.md
@@ -1,6 +1,6 @@
 0.0.5 / 2013-05-20
 ==================
-  * Added config file and command line support
+  * Added config file, command line support, more
   
   0.0.4 / 2013-05-20
 ==================